Split up codec hierarchy into `(de|en)coding::{format, framing}`
