+++ src/codecs/decoding/format/json.rs
@@ -1,19 +1,12 @@
 use std::convert::TryInto;
 
-use bytes::{BufMut, Bytes};
+use bytes::Bytes;
 use chrono::Utc;
 use serde::{Deserialize, Serialize};
 use smallvec::{smallvec, SmallVec};
-use tokio_util::codec::Encoder;
 
-use crate::{
-    codecs::{
-        decoding::{BoxedDeserializer, Deserializer, DeserializerConfig},
-        encoding::{BoxedSerializer, SerializerConfig},
-    },
-    config::log_schema,
-    event::Event,
-};
+use super::{BoxedDeserializer, Deserializer, DeserializerConfig};
+use crate::{config::log_schema, event::Event};
 
 /// Config used to build a `JsonDeserializer`.
 #[derive(Debug, Clone, Default, Deserialize, Serialize)]
@@ -84,55 +77,10 @@ impl From<&JsonDeserializerConfig> for JsonDeserializer {
     }
 }
 
-/// Config used to build a `JsonSerializer`.
-#[derive(Debug, Clone, Default, Deserialize, Serialize)]
-pub struct JsonSerializerConfig;
-
-impl JsonSerializerConfig {
-    /// Creates a new `JsonSerializerConfig`.
-    pub const fn new() -> Self {
-        Self
-    }
-}
-
-#[typetag::serde(name = "json")]
-impl SerializerConfig for JsonSerializerConfig {
-    fn build(&self) -> crate::Result<BoxedSerializer> {
-        Ok(Box::new(JsonSerializer))
-    }
-}
-
-/// Serializer that converts an `Event` to bytes using the JSON format.
-#[derive(Debug, Clone)]
-pub struct JsonSerializer;
-
-impl JsonSerializer {
-    /// Creates a new `JsonSerializer`.
-    pub const fn new() -> Self {
-        Self
-    }
-}
-
-impl Encoder<Event> for JsonSerializer {
-    type Error = crate::Error;
-
-    fn encode(&mut self, event: Event, buffer: &mut bytes::BytesMut) -> Result<(), Self::Error> {
-        let writer = buffer.writer();
-        match event {
-            Event::Log(log) => serde_json::to_writer(writer, &log),
-            Event::Metric(metric) => serde_json::to_writer(writer, &metric),
-        }
-        .map_err(Into::into)
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
     use crate::config::log_schema;
-    use crate::event::Value;
-    use bytes::BytesMut;
-    use vector_common::btreemap;
 
     #[test]
     fn deserialize_json() {
@@ -178,20 +126,7 @@ mod tests {
     }
 
     #[test]
-    fn serialize_json() {
-        let event = Event::from(btreemap! {
-            "foo" => Value::from("bar")
-        });
-        let mut serializer = JsonSerializer::new();
-        let mut bytes = BytesMut::new();
-
-        serializer.encode(event, &mut bytes).unwrap();
-
-        assert_eq!(bytes.freeze(), r#"{"foo":"bar"}"#);
-    }
-
-    #[test]
-    fn skip_empty() {
+    fn deserialize_skip_empty() {
         let input = Bytes::from("");
         let deserializer = JsonDeserializer::new();
 
@@ -202,7 +137,7 @@ mod tests {
     }
 
     #[test]
-    fn error_invalid_json() {
+    fn deserialize_error_invalid_json() {
         let input = Bytes::from("{ foo");
         let deserializer = JsonDeserializer::new();
