+++ src/codecs/decoding/framing/length_delimited.rs
@@ -2,7 +2,7 @@ use bytes::{Bytes, BytesMut};
 use serde::{Deserialize, Serialize};
 use tokio_util::codec::Decoder;
 
-use crate::codecs::decoding::{BoxedFramer, BoxedFramingError, FramingConfig};
+use super::{BoxedFramer, BoxedFramingError, FramingConfig};
 
 /// Config used to build a `LengthDelimitedDecoder`.
 #[derive(Debug, Clone, Default, Deserialize, Serialize)]
