+++ src/codecs/decoding/framing/newline_delimited.rs
@@ -1,8 +1,8 @@
 use bytes::{Bytes, BytesMut};
 use serde::{Deserialize, Serialize};
-use tokio_util::codec::{Decoder, Encoder};
+use tokio_util::codec::Decoder;
 
-use crate::codecs::{decoding, encoding, CharacterDelimitedDecoder, CharacterDelimitedEncoder};
+use super::{BoxedFramer, BoxedFramingError, CharacterDelimitedDecoder, FramingConfig};
 
 /// Config used to build a `NewlineDelimitedDecoder`.
 #[derive(Debug, Clone, Default, Deserialize, Serialize, PartialEq)]
@@ -49,8 +49,8 @@ impl NewlineDelimitedDecoderConfig {
 }
 
 #[typetag::serde(name = "newline_delimited")]
-impl decoding::FramingConfig for NewlineDelimitedDecoderConfig {
-    fn build(&self) -> crate::Result<decoding::BoxedFramer> {
+impl FramingConfig for NewlineDelimitedDecoderConfig {
+    fn build(&self) -> crate::Result<BoxedFramer> {
         if let Some(max_length) = self.newline_delimited.max_length {
             Ok(Box::new(NewlineDelimitedDecoder::new_with_max_length(
                 max_length,
@@ -89,7 +89,7 @@ impl Default for NewlineDelimitedDecoder {
 
 impl Decoder for NewlineDelimitedDecoder {
     type Item = Bytes;
-    type Error = decoding::BoxedFramingError;
+    type Error = BoxedFramingError;
 
     fn decode(&mut self, src: &mut BytesMut) -> Result<Option<Self::Item>, Self::Error> {
         self.0.decode(src)
@@ -97,43 +97,6 @@ impl Decoder for NewlineDelimitedDecoder {
 
     fn decode_eof(&mut self, src: &mut BytesMut) -> Result<Option<Self::Item>, Self::Error> {
         self.0.decode_eof(src)
-    }
-}
-
-/// Config used to build a `NewlineDelimitedEncoder`.
-#[derive(Debug, Clone, Default, Deserialize, Serialize, PartialEq)]
-pub struct NewlineDelimitedEncoderConfig;
-
-impl NewlineDelimitedEncoderConfig {
-    /// Creates a new `NewlineDelimitedEncoderConfig`.
-    pub fn new() -> Self {
-        Default::default()
-    }
-}
-
-#[typetag::serde(name = "newline_delimited")]
-impl encoding::FramingConfig for NewlineDelimitedEncoderConfig {
-    fn build(&self) -> crate::Result<encoding::BoxedFramer> {
-        Ok(Box::new(NewlineDelimitedEncoder::new()))
-    }
-}
-
-/// A codec for handling bytes that are delimited by (a) newline(s).
-#[derive(Debug, Clone)]
-pub struct NewlineDelimitedEncoder(CharacterDelimitedEncoder);
-
-impl NewlineDelimitedEncoder {
-    /// Creates a new `NewlineDelimitedEncoder`.
-    pub const fn new() -> Self {
-        Self(CharacterDelimitedEncoder::new(b'\n'))
-    }
-}
-
-impl Encoder<()> for NewlineDelimitedEncoder {
-    type Error = encoding::BoxedFramingError;
-
-    fn encode(&mut self, _: (), dst: &mut BytesMut) -> Result<(), Self::Error> {
-        self.0.encode((), dst)
     }
 }
