+++ src/codecs/decoding/framing/octet_counting.rs
@@ -4,7 +4,7 @@ use bytes::{Buf, Bytes, BytesMut};
 use serde::{Deserialize, Serialize};
 use tokio_util::codec::{LinesCodec, LinesCodecError};
 
-use crate::codecs::decoding::{BoxedFramer, BoxedFramingError, FramingConfig};
+use super::{BoxedFramer, BoxedFramingError, FramingConfig};
 
 /// Config used to build a `OctetCountingDecoder`.
 #[derive(Debug, Clone, Default, Deserialize, Serialize)]
