+++ src/codecs/decoding/mod.rs
@@ -1,16 +1,28 @@
 //! A collection of support structures that are used in the process of decoding
 //! bytes into events.
 
-use std::fmt::Debug;
+pub mod format;
+pub mod framing;
+
+pub use format::{
+    BoxedDeserializer, BytesDeserializer, BytesDeserializerConfig, Deserializer,
+    DeserializerConfig, JsonDeserializer, JsonDeserializerConfig,
+};
+#[cfg(feature = "sources-syslog")]
+pub use format::{SyslogDeserializer, SyslogDeserializerConfig};
+pub use framing::{
+    BoxedFramer, BoxedFramingError, BytesDecoder, BytesDecoderConfig, CharacterDelimitedDecoder,
+    CharacterDelimitedDecoderConfig, Framer, FramingConfig, FramingError, LengthDelimitedDecoder,
+    LengthDelimitedDecoderConfig, NewlineDelimitedDecoder, NewlineDelimitedDecoderConfig,
+    OctetCountingDecoder, OctetCountingDecoderConfig,
+};
 
 use bytes::{Bytes, BytesMut};
-use dyn_clone::DynClone;
 use serde::{Deserialize, Serialize};
 use smallvec::SmallVec;
-use tokio_util::codec::LinesCodecError;
+use std::fmt::Debug;
 
 use crate::{
-    codecs::{BytesDeserializer, NewlineDelimitedDecoder},
     event::Event,
     internal_events::{DecoderDeserializeFailed, DecoderFramingFailed},
     sources::util::StreamDecodingError,
@@ -52,109 +64,6 @@ impl StreamDecodingError for Error {
         }
     }
 }
-
-/// An error that occurred while producing byte frames from a byte stream / byte
-/// message.
-///
-/// It requires conformance to `TcpError` so that we can determine whether the
-/// error is recoverable or if trying to continue will lead to hanging up the
-/// TCP source indefinitely.
-pub trait FramingError: std::error::Error + StreamDecodingError + Send + Sync {}
-
-impl std::error::Error for BoxedFramingError {}
-
-impl FramingError for std::io::Error {}
-
-impl FramingError for LinesCodecError {}
-
-impl From<std::io::Error> for BoxedFramingError {
-    fn from(error: std::io::Error) -> Self {
-        Box::new(error)
-    }
-}
-
-impl From<LinesCodecError> for BoxedFramingError {
-    fn from(error: LinesCodecError) -> Self {
-        Box::new(error)
-    }
-}
-
-/// A `Box` containing a `FramingError`.
-pub type BoxedFramingError = Box<dyn FramingError>;
-
-impl StreamDecodingError for BoxedFramingError {
-    fn can_continue(&self) -> bool {
-        self.as_ref().can_continue()
-    }
-}
-
-/// Produce byte frames from a byte stream / byte message.
-pub trait Framer:
-    tokio_util::codec::Decoder<Item = Bytes, Error = BoxedFramingError> + DynClone + Debug + Send + Sync
-{
-}
-
-/// Default implementation for `Framer`s that implement
-/// `tokio_util::codec::Decoder`.
-impl<Decoder> Framer for Decoder where
-    Decoder: tokio_util::codec::Decoder<Item = Bytes, Error = BoxedFramingError>
-        + Clone
-        + Debug
-        + Send
-        + Sync
-{
-}
-
-dyn_clone::clone_trait_object!(Framer);
-
-/// A `Box` containing a `Framer`.
-pub type BoxedFramer = Box<dyn Framer>;
-
-/// Define options for a framer and build it from the config object.
-///
-/// Implementors must annotate the struct with `#[typetag::serde(name = "...")]`
-/// to define which value should be read from the `method` key to select their
-/// implementation.
-#[typetag::serde(tag = "method")]
-pub trait FramingConfig: Debug + DynClone + Send + Sync {
-    /// Builds a framer from this configuration.
-    ///
-    /// Fails if the configuration is invalid.
-    fn build(&self) -> crate::Result<BoxedFramer>;
-}
-
-dyn_clone::clone_trait_object!(FramingConfig);
-
-/// Parse structured events from bytes.
-pub trait Deserializer: DynClone + Debug + Send + Sync {
-    /// Parses structured events from bytes.
-    ///
-    /// It returns a `SmallVec` rather than an `Event` directly, since one byte
-    /// frame can potentially hold multiple events, e.g. when parsing a JSON
-    /// array. However, we optimize the most common case of emitting one event
-    /// by not requiring heap allocations for it.
-    fn parse(&self, bytes: Bytes) -> crate::Result<SmallVec<[Event; 1]>>;
-}
-
-dyn_clone::clone_trait_object!(Deserializer);
-
-/// A `Box` containing a `Deserializer`.
-pub type BoxedDeserializer = Box<dyn Deserializer>;
-
-/// Define options for a deserializer and build it from the config object.
-///
-/// Implementors must annotate the struct with `#[typetag::serde(name = "...")]`
-/// to define which value should be read from the `codec` key to select their
-/// implementation.
-#[typetag::serde(tag = "codec")]
-pub trait DeserializerConfig: Debug + DynClone + Send + Sync {
-    /// Builds a deserializer from this configuration.
-    ///
-    /// Fails if the configuration is invalid.
-    fn build(&self) -> crate::Result<BoxedDeserializer>;
-}
-
-dyn_clone::clone_trait_object!(DeserializerConfig);
 
 #[derive(Debug, Clone)]
 /// A decoder that can decode structured events from a byte stream / byte
