+++ src/codecs/encoding/format/json.rs
@@ -1,0 +1,69 @@
+use bytes::{BufMut, BytesMut};
+use serde::{Deserialize, Serialize};
+use tokio_util::codec::Encoder;
+
+use super::{BoxedSerializer, SerializerConfig};
+use crate::event::Event;
+
+/// Config used to build a `JsonSerializer`.
+#[derive(Debug, Clone, Default, Deserialize, Serialize)]
+pub struct JsonSerializerConfig;
+
+impl JsonSerializerConfig {
+    /// Creates a new `JsonSerializerConfig`.
+    pub const fn new() -> Self {
+        Self
+    }
+}
+
+#[typetag::serde(name = "json")]
+impl SerializerConfig for JsonSerializerConfig {
+    fn build(&self) -> crate::Result<BoxedSerializer> {
+        Ok(Box::new(JsonSerializer))
+    }
+}
+
+/// Serializer that converts an `Event` to bytes using the JSON format.
+#[derive(Debug, Clone)]
+pub struct JsonSerializer;
+
+impl JsonSerializer {
+    /// Creates a new `JsonSerializer`.
+    pub const fn new() -> Self {
+        Self
+    }
+}
+
+impl Encoder<Event> for JsonSerializer {
+    type Error = crate::Error;
+
+    fn encode(&mut self, event: Event, buffer: &mut BytesMut) -> Result<(), Self::Error> {
+        let writer = buffer.writer();
+        match event {
+            Event::Log(log) => serde_json::to_writer(writer, &log),
+            Event::Metric(metric) => serde_json::to_writer(writer, &metric),
+        }
+        .map_err(Into::into)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::event::Value;
+    use bytes::BytesMut;
+    use vector_common::btreemap;
+
+    #[test]
+    fn serialize_json() {
+        let event = Event::from(btreemap! {
+            "foo" => Value::from("bar")
+        });
+        let mut serializer = JsonSerializer::new();
+        let mut bytes = BytesMut::new();
+
+        serializer.encode(event, &mut bytes).unwrap();
+
+        assert_eq!(bytes.freeze(), r#"{"foo":"bar"}"#);
+    }
+}
